# mnemonic/internal/http/ui/openapi.yaml
openapi: 3.0.3
info:
  title: Skillgrid Mnemonic HTTP API
  description: |-
    Persistent memory, code index, and web research cache for AI agents.

    Write routes (POST) optionally require an `Authorization: Bearer <token>`
    header when the server is started with `SKILLGRID_HTTP_TOKEN` set.
    Most routes take a `project` query parameter naming the project store
    (see `GET /projects` for available IDs).
  version: "0.1.0"
servers:
  - url: /
...
paths:
...
  /policy:
    get:
      tags: [session]
      operationId: policyGet
      summary: The merged pre-tool policy (read-only)
      parameters:
        - { name: project, in: query, required: true, schema: { type: string } }
        - { name: directory, in: query, schema: { type: string } }
      responses:
        "200":
          description: Enabled flag, rules in evaluation order, and source files
          content:
            application/json:
              schema:
                type: object
                properties:
                  project: { type: string }
                  enabled: { type: boolean }
                  rules: { type: array, items: { type: object } }
                  files: { type: array, items: { type: string } }
                  repoFile: { type: string }
                  error: { type: string }
...
  /context:
    get:
      tags: [memory]
      operationId: context
      summary: Recent session context
      parameters:
        - name: project
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Recent sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sessions"
        "400":
          $ref: "#/components/responses/Bad"
        "500":
          $ref: "#/components/responses/Internal"
...
  /tracker/providers:
    get:
      tags: [tracker]
      operationId: trackerProviders
      summary: Active tracker provider + whether its CLI is connected
      description: |-
        Reports the resolved provider (never guessed — 501 when unresolvable) and
        whether a CLI-backed provider's binary is on PATH. File-based Backlog.md is
        always connected:true. The UI uses connected:false to show a "not connected"
        empty state instead of a perpetual spinner.
      responses:
        "200":
          description: Provider status
          content:
            application/json:
              schema:
                type: object
                required: [provider, connected]
                properties:
                  provider:
                    type: string
                  connected:
                    type: boolean
                  reason:
                    type: string
                  cli:
                    type: string
              example:
                provider: backlogmd
                connected: true
                reason: ""
                cli: ""
        "501":
          $ref: "#/components/responses/Bad"
  /tracker/config:
    get:
      tags: [tracker]
      operationId: trackerConfig
      summary: Active tracker provider with its statuses, types, and priorities
      description: |-
        Resolved from SKILLGRID_TRACKER, docs/skillgrid/agents/issue-tracker.md,
        then config.yaml (default backlogmd). 501 when unresolvable, 503 when the
        provider CLI is missing, 502 on CLI failure.
      responses:
        "200":
          description: Provider config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackerConfig"
              example:
                provider: backlogmd
                statuses: [needs-triage, ready-for-agent, done]
                types: [feature, bug, chore]
                priorities: [high, medium, low]
                schema_version: "1"
        "501":
          $ref: "#/components/responses/Bad"
        "502":
          $ref: "#/components/responses/Internal"
        "503":
          $ref: "#/components/responses/Internal"
  /tracker/tasks:
    get:
      tags: [tracker]
      operationId: trackerList
      summary: List tracker items as normalized DTOs
      responses:
        "200":
          description: Normalized item list with provider name
          content:
            application/json:
              schema:
                type: object
                required: [tasks, provider]
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/TrackerItem"
                  provider:
                    type: string
              example:
                provider: backlogmd
                tasks:
                  - id: TASK-001
                    title: Fix login
                    status: needs-triage
                    type: feature
                    priority: high
                    ac_completed: 1
                    ac_total: 4
                    is_ready: true
                    provider: backlogmd
        "501":
          $ref: "#/components/responses/Bad"
        "502":
          $ref: "#/components/responses/Internal"
        "503":
          $ref: "#/components/responses/Internal"
  /tracker/tasks/{id}:
    get:
      tags: [tracker]
      operationId: trackerGet
      summary: One tracker item as a normalized DTO
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Normalized item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackerItem"
        "404":
          $ref: "#/components/responses/Bad"
        "501":
          $ref: "#/components/responses/Bad"
        "502":
          $ref: "#/components/responses/Internal"
        "503":
          $ref: "#/components/responses/Internal"
    post:
      tags: [tracker]
      operationId: trackerSetStatus
      summary: Change an item's status (provider-native, write-gated)
      description: |-
        Backlog.md runs `task edit -s`; GitHub/GitLab close/reopen; Jira runs
        `issue move`. Requires the bearer token when SKILLGRID_HTTP_TOKEN is set.
        400 on a status outside the provider set, 501 where the provider has no
        equivalent transition.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
            example:
              status: ready-for-agent
      responses:
        "200":
          description: Updated normalized item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackerItem"
        "400":
          $ref: "#/components/responses/Bad"
        "401":
          $ref: "#/components/responses/Bad"
        "404":
          $ref: "#/components/responses/Bad"
        "501":
          $ref: "#/components/responses/Bad"
        "502":
          $ref: "#/components/responses/Internal"
        "503":
          $ref: "#/components/responses/Internal"
    patch:
      tags: [tracker]
      operationId: trackerPatch
      summary: Update an item's status (the Phase 2 write path, write-gated)
      description: |-
        PATCH /tracker/tasks/{id} with {"status": ...} → SetStatus. Same semantics
        as POST .../status but using the HTTP PATCH verb. Requires the bearer token
        when SKILLGRID_HTTP_TOKEN is set. 400 on a status outside the provider set,
        501 where the provider has no equivalent transition.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
            example:
              status: ready-for-agent
      responses:
        "200":
          description: Updated normalized item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackerItem"
        "400":
          $ref: "#/components/responses/Bad"
        "401":
          $ref: "#/components/responses/Bad"
        "404":
          $ref: "#/components/responses/Bad"
        "501":
          $ref: "#/components/responses/Bad"
        "502":
          $ref: "#/components/responses/Internal"
        "503":
          $ref: "#/components/responses/Internal"
  /tracker/tasks/{id}/deps:
    get:
      tags: [tracker]
      operationId: trackerDeps
      summary: Dependency edges for one task (the Kanban dependency mini-graph)
      description: |-
        Backlog.md computes deps from the frontmatter `dependencies` list across all
        tasks. GitHub/GitLab/Jira expose no native edges here, so they return empty
        (the board still renders). 404 for an unknown task id.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dependency edges
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackerDeps"
              example:
                task_id: TASK-001
                deps_in: [TASK-002]
                deps_out: []
        "404":
          $ref: "#/components/responses/Bad"
        "501":
          $ref: "#/components/responses/Bad"
...
  /backlog/config:
    get:
      tags: [tracker]
      operationId: backlogConfigAlias
      summary: Alias of GET /tracker/config on Backlog.md repos (else 501)
      responses:
        "200":
          description: Provider config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackerConfig"
        "501":
          $ref: "#/components/responses/Bad"
  /backlog/tasks:
    get:
      tags: [tracker]
      operationId: backlogListAlias
      summary: Alias of GET /tracker/tasks on Backlog.md repos (else 501)
      responses:
        "200":
          description: Normalized item list with provider name
          content:
            application/json:
              schema:
                type: object
                required: [tasks, provider]
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/TrackerItem"
                  provider:
                    type: string
        "501":
          $ref: "#/components/responses/Bad"
  /backlog/tasks/{id}:
    get:
      tags: [tracker]
      operationId: backlogGetAlias
      summary: Alias of GET /tracker/tasks/{id} on Backlog.md repos (else 501)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Normalized item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackerItem"
        "404":
          $ref: "#/components/responses/Bad"
        "501":
          $ref: "#/components/responses/Bad"
...
components:
  responses:
    Bad:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
    Internal:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
...
  schemas:
...
    HTTPError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    TrackerItem:
      type: object
      required: [id, title, status, board, provider]
      properties:
        id:
          type: string
          description: Provider-native identifier (TASK-001, 12, PROJ-1)
        title:
          type: string
        status:
          type: string
          description: Provider-native status
        status_detail:
          type: string
        description:
          type: string
          description: Item body (Backlog.md description, GitHub body, GitLab description)
        board:
          type: string
          description: Canonical kanban column (todo|in_progress|blocked|done) for one consistent board
          enum: [todo, in_progress, blocked, done]
        type:
          type: string
        priority:
          type: string
        assignees:
          type: array
          items:
            type: string
        labels:
          type: array
          items:
            type: string
        doc_refs:
          type: array
          description: Linked document paths (used by the Docs entry)
          items:
            type: string
        dependencies:
          type: array
          description: Task ids this task depends on (edges out — the Kanban dependency mini-graph)
          items:
            type: string
        milestone:
          type: string
          description: Optional milestone label
        parent:
          type: string
          description: Optional parent task id (subtask linkage)
        due_date:
          type: string
          description: Optional due date (day, no time)
        ac_completed:
          type: integer
        ac_total:
          type: integer
        is_ready:
          type: boolean
          nullable: true
        created_at:
          type: string
        updated_at:
          type: string
        provider:
          type: string
    TrackerDeps:
      type: object
      required: [task_id]
      properties:
        task_id:
          type: string
        deps_in:
          type: array
          description: Tasks that depend on this one (its dependents)
          items:
            type: string
        deps_out:
          type: array
          description: Tasks this one depends on (its blockers)
          items:
            type: string
    TrackerConfig:
      type: object
      required: [provider, statuses]
      properties:
        provider:
          type: string
        statuses:
          type: array
          items:
            type: string
        types:
          type: array
          items:
            type: string
        priorities:
          type: array
          items:
            type: string
        version:
          type: string
        schema_version:
          type: string
        repo:
          type: string
        project_key:
          type: string
...
    Sessions:
      type: object
      required: [sessions]
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/Session"
...
    Session:
      type: object
      properties:
        id:
          type: string
        project:
          type: string
        directory:
          type: string
        title:
          type: string
          nullable: true
          description: Human-readable session name (optional); shown in the dashboard session list
        started_at:
          type: string
        ended_at:
          type: string
        summary:
          type: string
        status:
          type: string
...